VSCode: drop dead deps and unify workspace-file collection
Remove the unused @vscode/test-web devDependency (referenced by no script,
test, or workflow) and refresh the lockfile. Drop the two dead .vscodeignore
entries .oxfmtrc.json/.oxlintrc.json (neither file exists).

Collapse the two independent collectWorkspaceFiles implementations (extension.ts
and debug-adapter.ts) into a single shared helper in workspace-paths.ts that
also owns FILE_GLOB. The debug adapter's extra launch-program resolution is now
an optional parameter of that helper. Behavior-preserving.
