+++ vscode/package.json
@@ -133,7 +133,6 @@
   },
   "devDependencies": {
     "@types/vscode": "~1.100.0",
-    "@vscode/test-web": "^0.0.81",
     "@vscode/vsce": "^3.9.2",
     "esbuild": "^0.28.0",
     "oxfmt": "^0.57.0",
